edit/update getting working

+++ server/controllers/BraceletController.js
@@ -45,22 +45,18 @@ module.exports = {
             .catch(err => res.status(400).json(err))
     },
     deleteBracelet: (req, res) => {
-        // Bracelet.findByIdAndDelete({_id: req.params.id})
-        //     .then(deletedBracelet =>{
-        //         console.log("Bracelet was deleted")
-        //         res.json(deletedBracelet)
-        //     })
-        //     .catch(err => res.status(400).json(err))
-        
-            Bracelet.findById (req.params.id)
+            Bracelet.findByIdAndDelete (req.params.id)
                 .then(bracelet => {
                     // console.log(bracelet.image)
-                    fs.unlinkSync(bracelet.image)
-                    return bracelet.remove()
-                })
-                .then(() => {
-                    res.json({deleted:true})
-                })
-        
+                    if(bracelet.image)
+                        { try{
+                            fs.unlinkSync(bracelet.image)
+                        }
+                        catch (error){
+                        console.log('Error deleting image file: ', error);
+                    }
+                    return res.json({deleted: true})
+                }})
+                .catch(err => res.status(400).json(err));
     }
 };
